images

+++ app/Http/Controllers/Api/Admin/ScriptTypeController.php
@@ -218,7 +218,6 @@ class ScriptTypeController extends Controller
             'name' => $request['name'],
             'prompt_1' => $request['prompt_1'],
             'prompt_2' => $request['prompt_2'],
-            'icon' => $request['icon'],
             'description' => $request['description'],
             'presence_penalty' => $request['presence_penalty'],
             'frequency_penalty' => $request['frequency_penalty'],
@@ -231,17 +230,20 @@ class ScriptTypeController extends Controller
             'top_p' => $request['top_p'],
             'engine' => $request['engine'],
         ]);
-        
-        foreach($request['script_type_presets'] as $preset){
-            $model->presets()->delete();
-            $model->presets()->create($preset);
+
+        if($request['script_type_presets']){
+            foreach($request['script_type_presets'] as $preset){
+                $model->presets()->delete();
+                $model->presets()->create($preset);
+            }
         }
 
         if($request->has('icon')){
             $image = $request['icon'];
             if (gettype($image) != "integer") {
                 $path = $this->uploadImage($image, "icon");
-                $model->avatar()->create([
+
+                $model->iconImage()->create([
                     'file_path' => $path,
                 ]);
             } else {
